[#2224] Move out metadata utils from workshops.util
